refactor: split checker declaration metadata
